[SQONE-519]: Rename masthead component to site-header for consistency. Move document header css to that component. Consolidate css partials for ease of use.
