+++ wp-content/themes/core/components/document/css/header.pcss
@@ -1,7 +1,16 @@
 /* -----------------------------------------------------------------------------
  *
- * Accessibility Skip Link
+ * Component: Document Header
  *
+ * ----------------------------------------------------------------------------- */
+
+/* WP Admin Bar */
+#wpadminbar {
+	position: fixed !important;
+}
+
+/* -----------------------------------------------------------------------------
+ * Accessibility Skip Link
  * ----------------------------------------------------------------------------- */
 
 .a11y-skip-link {
