+++ wp-content/themes/core/components/document/index.pcss
@@ -1,12 +1,10 @@
 /* -----------------------------------------------------------------------------
  *
- * Component: Header
+ * Component: Document Header
  *
  * This file is just a clearing-house, see the css directory
  * and edit the source files found there.
  *
  * ----------------------------------------------------------------------------- */
 
-@import "css/skip-link.pcss";
 @import "css/header.pcss";
-@import "css/logo.pcss";
