--- wp-content/themes/core/components/header/css/header.pcss
@@ -1,8 +1,0 @@
-.site-header {
-
-}
-
-/* WP Admin Bar */
-#wpadminbar {
-	position: fixed !important;
-}
