--- wp-content/themes/core/components/header/css/logo.pcss
@@ -1,3 +1,0 @@
-.logo {
-
-}
