+++ wp-content/themes/core/components/site_header/css/site-header.pcss
@@ -1,0 +1,9 @@
+/* -----------------------------------------------------------------------------
+ *
+ * Component: Site Header
+ *
+ * ----------------------------------------------------------------------------- */
+
+.site-header {
+
+}
