+++ wp-content/themes/core/components/site_header/index.pcss
@@ -1,0 +1,10 @@
+/* -----------------------------------------------------------------------------
+ *
+ * Component: Site Header
+ *
+ * This file is just a clearing-house, see the css directory
+ * and edit the source files found there.
+ *
+ * ----------------------------------------------------------------------------- */
+
+@import "css/site-header.pcss";
